caveats: making plist_caveats macOS-only
