+++ Library/Homebrew/extend/os/caveats.rb
@@ -1,0 +1,1 @@
+require "extend/os/mac/caveats" if OS.mac?
