--- Library/Homebrew/extend/os/linux/caveats.rb
@@ -1,3 +1,0 @@
-class Caveats
-  def plist_caveats; end
-end
